% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_periods.R
\name{add_periods}
\alias{add_periods}
\title{Add new fiscal information fields to a dataframe}
\usage{
add_periods(df, date_col)
}
\arguments{
\item{df}{dataframe you want add information to}

\item{date_col}{field on which all calculations will be made}
}
\description{
This function will return a dataframe with fiscal period information added. It will only accept valid date fields (date, POSIXct, POSIXlt)
}
\examples{
add_periods(myydataframe, "createddate_col")
will return mydataframe with the following new columns:
createddate_col_period
createddate_col_qtr
createddate_col_year
createddate_col_year_qtr
createddate_col_month_end
}
