% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_data.R
\name{filt_data}
\alias{filt_data}
\title{Filter the Data}
\usage{
filt_data(variables_list, max_times, codes_list)
}
\arguments{
\item{variables_list}{the variables that we want to filter.
Let df be a dataframe, the variable_list will be df[,questions_to_filter].}

\item{max_times}{The number of maximum times that a code can appear.}

\item{codes_list}{The array of codes that we want to filter.}
}
\value{
returns an array with boolean variables that remark which columns are
NOT to be removed
}
\description{
This function is used to subset your dataframe if you want to remove some rows
which some speciefic code appears more than max_times.
}
\details{
For example: If you are doing a poll and a person is saying "DK/NA" multiple
times. It is normal to invalid this opinion. With this function you can subset
your dataset easily.
}
\examples{
p15_1 <- c(0, 2, 3, 99, 99, 99, 2)
p15_2 <- c(1, 2, 5, 99, 4, 52, 2)
data_frame <- data.frame(p15_1, p15_2)
data_frame_output <-data_frame[ filt_data(
  variables_list = data_frame[,c ("p15_1", "p15_2")],
  max_times =  1,
  codes_list =   99
),]

}
