% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dates.R
\name{fix_dates}
\alias{fix_dates}
\title{Fix a character field to date}
\usage{
fix_dates(df, format = "\%Y-\%m-\%d \%H:\%M:\%S", date_search = "date")
}
\arguments{
\item{df}{dataframe you want add information to}

\item{format}{formt that date is currently in so we can parse it}

\item{date_search}{character string to search for within the names of each column; Any that match will be changed to a date format}
}
\description{
This function will return a dataframe with all
}
\examples{
fix_dates(myydataframe)
}
