% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPSS.R
\name{FRE}
\alias{FRE}
\title{FRE
This function replicates the FREQUENCIES function of SPSS. It accepts a
dataframe and returns a table in the format of a dataframe with the number of
cases and the propotion.}
\usage{
FRE(df, variable)
}
\arguments{
\item{df}{the Dataframe that you introduced from the function haven::read_sav()}

\item{variable}{the variable we want to see the frequencies}
}
\value{
returns a dataframe whose columns are the rows of the tables.
}
\description{
FRE
This function replicates the FREQUENCIES function of SPSS. It accepts a
dataframe and returns a table in the format of a dataframe with the number of
cases and the propotion.
}
\examples{
p15_1 <- c(0, 2, 3, 99, 99, 99, 2)
p15_2 <- c(1, 2, 5, 3, 4, 52, 2)
data_frame <- data.frame(p15_1, p15_2)
FRE(data_frame,"p15_1")

}
