% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPSS.R
\name{frec}
\alias{frec}
\title{frec}
\usage{
frec(x, w)
}
\arguments{
\item{x}{is the vector that counts how many times an event has occured}

\item{w}{is a vector of the weight.}
}
\value{
returns a vector with all of the information (it's a cbind of the table
and the prop.table).
}
\description{
This function returns a table which first column is the frequencies of the event
and the second column is the percentatge of this frequencies.
}
\examples{
p15_1 <- c(0, 2, 3, 99, 99, 99, 2)
p15_2 <- c(1, 2, 5, 3, 4, 52, 2)
weight <- c(200,400,50,200,100,300,90)
data_frame <- data.frame(p15_1, p15_2)
frec(data_frame[,"p15_1"],weight)

}
