% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondf.R
\name{mondf}
\alias{mondf}
\title{Get the difference in months between two dates}
\usage{
mondf(d1, d2)
}
\arguments{
\item{d1}{Date From}

\item{d2}{Date To}
}
\description{
This function will retunr the number of months between two dates
}
\examples{
mondf()
}
