% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_data.R
\name{filt_data}
\alias{filt_data}
\title{Filter the Data}
\usage{
filt_data(df, variables_list, max_times, codes_list)
}
\arguments{
\item{df}{the Dataframe that you introduced from the function haven::read_sav()}

\item{max_times}{The number of maximum times that a code can appear.}

\item{codes_list}{The array of codes that we want to filter.}

\item{variable_list}{the variables that we want to filter.}
}
\value{
returns the df without these rows.
}
\description{
This function removes the answers that have said one specific code more times
that max_times. For example, in a poll we descart people who said 'Ns/Nr' more
than 12 times.
}
\examples{
filt_data(df=data_frame, variable_list= c("p15_1", "p15_2"), max_times=7, codes_list=c(97,98,99))
}
